# src/api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Welcome to MalURLs, an API to classify URLs",
    "version": "1.0.0",
    "title": "MalURLs"
  },
  "tags": [
    {
      "name": "Products API",
      "description": "Test API"
    }
  ],
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "schemas": {
      "TodoList": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Id": {
        "type": "integer",
        "properties": {
          "name": {
            "type": "integer"
          }
        }
      },
      "update": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "tags": ["Main"],
        "summary": "Get main page details",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "main_page_details": "Your main page details here"
                }
              }
            }
          }
        }
      }
    },
    "/get_features": {
      "post": {
        "tags": ["Feature Operations"],
        "summary": "Get features for a URL",
        "requestBody": {
          "description": "URL details",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "url": "https://example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "url_features": {},
                  "timestamp": "2023-01-01T00:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "URL must be a non-empty string"
                }
              }
            }
          }
        }
      }
    },
    "/models": {
      "get": {
        "tags": ["Model Operations"],
        "summary": "Get available models",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "models_available": ["base_rf", "tuned_rf"]
                }
              }
            }
          }
        }
      }
    },
    "/scan": {
      "post": {
        "tags": ["Scan Operations"],
        "summary": "Scan a URL with a selected model",
        "requestBody": {
          "description": "URL and model details",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "url": "https://example.com",
                "model": "base_rf"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "prediction": "malicious",
                  "model_used": "base_rf",
                  "timestamp": "2023-01-01T00:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "URL must be a non-empty string"
                }
              }
            }
          }
        }
      }
    },
    "/scan_all": {
      "post": {
        "tags": ["Scan Operations"],
        "summary": "Scan a URL with all models",
        "requestBody": {
          "description": "URL details",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "url": "https://example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "Base random forest prediction": "malicious",
                  "Tuned random forest prediction": "benign",
                  "timestamp": "2023-01-01T00:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "URL must be a non-empty string"
                }
              }
            }
          }
        }
      }
    }
  }
}